Preload singletons When migrating content, items can depend on records from other tables. For cases when there's a cylcical dependency, this is even more complicated. That's why we first create the records without any data when migrating content, just the primary key. After all records are created, we just need to update with the actual data.
This is the right approach but it's only working for collections with multiple records.
When a new singleton collection is migrated, we need to ensure that a record is there.
This could be something that Directus needs to handle when creating the collection via schema endpoint, but as of right now we should do that here so it migrate the content as expected.