Migration Bundle: Fix duplicates (#260)
* Fix duplicates
When migrating a large amount of data, it is not guaranteed that a result does not appear in different pages, when using pagination.

For that reason we need to also include "sort" parameter when using "page"

* Preload singletons
When migrating content, items can depend on records from other tables. For cases when there's a cylcical dependency, this is even more complicated.
That's why we first create the records without any data when migrating content, just the primary key.
After all records are created, we just need to update with the actual data.

This is the right approach but it's only working for collections with multiple records.
When a new singleton collection is migrated, we need to ensure that a record is there.
This could be something that Directus needs to handle when creating the collection via schema endpoint, but as of right now we should do that here so it migrate the content as expected.

* Lint

* Revert "Preload singletons"

This reverts commit 3beebf7298d1dc9e21bd30a852b39953dde824e6.

---------

Co-authored-by: ian <licitdev@gmail.com>